// src/screens/institution/signup/types.ts
export interface InstitutionApplicationValues {
  institutionName: string;
  institutionType: string;
  country: string;
  state: string;
  city: string;
  institutionAddress: string;
  officialEmail: string;
  phone: string;
  repFullName: string;
  repEmail: string;
  repPhone: string;
  repPosition: string;
}